+++ grpc.gyp
@@ -761,6 +761,7 @@
         'src/core/ext/xds/file_watcher_certificate_provider_factory.cc',
         'src/core/ext/xds/xds_api.cc',
         'src/core/ext/xds/xds_bootstrap.cc',
+        'src/core/ext/xds/xds_bootstrap_grpc.cc',
         'src/core/ext/xds/xds_certificate_provider.cc',
         'src/core/ext/xds/xds_channel_stack_modifier.cc',
         'src/core/ext/xds/xds_client.cc',
