+++ src/core/ext/filters/client_channel/lb_policy/xds/cds.cc
@@ -39,7 +39,7 @@
 
 #include "src/core/ext/filters/client_channel/lb_policy/outlier_detection/outlier_detection.h"
 #include "src/core/ext/xds/certificate_provider_store.h"
-#include "src/core/ext/xds/xds_bootstrap.h"
+#include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_certificate_provider.h"
 #include "src/core/ext/xds/xds_client.h"
 #include "src/core/ext/xds/xds_client_grpc.h"
@@ -444,8 +444,8 @@ absl::StatusOr<bool> CdsLb::GenerateDiscoveryMechanismForCluster(
       break;
   }
   if (state.update->lrs_load_reporting_server.has_value()) {
-    mechanism["lrsLoadReportingServer"] =
-        state.update->lrs_load_reporting_server->ToJson();
+    mechanism["lrsLoadReportingServer"] = GrpcXdsBootstrap::XdsServerToJson(
+        *state.update->lrs_load_reporting_server);
   }
   discovery_mechanisms->emplace_back(std::move(mechanism));
   return true;
