+++ src/core/ext/filters/client_channel/lb_policy/xds/xds_cluster_impl.cc
@@ -44,6 +44,7 @@
 #include "src/core/ext/filters/client_channel/lb_policy/xds/xds.h"
 #include "src/core/ext/filters/client_channel/lb_policy/xds/xds_channel_args.h"
 #include "src/core/ext/xds/xds_bootstrap.h"
+#include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_client.h"
 #include "src/core/ext/xds/xds_client_grpc.h"
 #include "src/core/ext/xds/xds_client_stats.h"
@@ -761,7 +762,7 @@ class XdsClusterImplLbFactory : public LoadBalancingPolicyFactory {
             "field:lrsLoadReportingServer error:type should be object");
       } else {
         grpc_error_handle parser_error;
-        lrs_load_reporting_server = XdsBootstrap::XdsServer::Parse(
+        lrs_load_reporting_server = GrpcXdsBootstrap::XdsServerParse(
             it->second.object_value(), &parser_error);
         if (!GRPC_ERROR_IS_NONE(parser_error)) {
           errors.emplace_back(
