+++ src/core/ext/filters/client_channel/lb_policy/xds/xds_cluster_resolver.cc
@@ -47,6 +47,7 @@
 #include "src/core/ext/filters/client_channel/lb_policy/xds/xds_channel_args.h"
 #include "src/core/ext/filters/client_channel/resolver/fake/fake_resolver.h"
 #include "src/core/ext/xds/xds_bootstrap.h"
+#include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_client.h"
 #include "src/core/ext/xds/xds_client_grpc.h"
 #include "src/core/ext/xds/xds_client_stats.h"
@@ -925,7 +926,8 @@ XdsClusterResolverLb::CreateChildPolicyConfigLocked() {
       }
       if (discovery_config.lrs_load_reporting_server.has_value()) {
         xds_cluster_impl_config["lrsLoadReportingServer"] =
-            discovery_config.lrs_load_reporting_server->ToJson();
+            GrpcXdsBootstrap::XdsServerToJson(
+                *discovery_config.lrs_load_reporting_server);
       }
       Json locality_picking_policy;
       if (XdsOutlierDetectionEnabled()) {
@@ -1191,7 +1193,7 @@ class XdsClusterResolverLbFactory : public LoadBalancingPolicyFactory {
       } else {
         grpc_error_handle parse_error;
         discovery_mechanism->lrs_load_reporting_server.emplace(
-            XdsBootstrap::XdsServer::Parse(it->second, &parse_error));
+            GrpcXdsBootstrap::XdsServerParse(it->second, &parse_error));
         if (!GRPC_ERROR_IS_NONE(parse_error)) {
           error_list.push_back(GRPC_ERROR_CREATE_FROM_CPP_STRING(
               absl::StrCat("errors parsing lrs_load_reporting_server")));
