+++ test/core/xds/xds_bootstrap_test.cc
@@ -14,20 +14,14 @@
 // limitations under the License.
 //
 
-#include "src/core/ext/xds/xds_bootstrap.h"
-
-#include <regex>
-
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
 #include "absl/strings/numbers.h"
 #include "absl/strings/str_format.h"
 
-#include <grpc/grpc.h>
-#include <grpc/slice.h>
-
 #include "src/core/ext/xds/certificate_provider_registry.h"
+#include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_client_grpc.h"
 #include "src/core/lib/gpr/env.h"
 #include "src/core/lib/gpr/tmpfile.h"
@@ -121,7 +115,7 @@ TEST(XdsBootstrapTest, Basic) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
@@ -189,7 +183,7 @@ TEST(XdsBootstrapTest, ValidWithoutNode) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
@@ -211,7 +205,7 @@ TEST(XdsBootstrapTest, InsecureCreds) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
@@ -249,7 +243,7 @@ TEST(XdsBootstrapTest, GoogleDefaultCreds) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
@@ -270,7 +264,7 @@ TEST(XdsBootstrapTest, MissingChannelCreds) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(
@@ -292,7 +286,7 @@ TEST(XdsBootstrapTest, NoKnownChannelCreds) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -305,7 +299,7 @@ TEST(XdsBootstrapTest, MissingXdsServers) {
   auto json = Json::Parse("{}");
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -324,7 +318,7 @@ TEST(XdsBootstrapTest, TopFieldsWrongTypes) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -346,7 +340,7 @@ TEST(XdsBootstrapTest, XdsServerMissingServerUri) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(
@@ -371,7 +365,7 @@ TEST(XdsBootstrapTest, XdsServerUriAndCredsWrongTypes) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -402,7 +396,7 @@ TEST(XdsBootstrapTest, ChannelCredsFieldsWrongTypes) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(
@@ -430,7 +424,7 @@ TEST(XdsBootstrapTest, NodeFieldsWrongTypes) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -456,7 +450,7 @@ TEST(XdsBootstrapTest, LocalityFieldsWrongType) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -484,7 +478,7 @@ TEST(XdsBootstrapTest, CertificateProvidersElementWrongType) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -512,7 +506,7 @@ TEST(XdsBootstrapTest, CertificateProvidersPluginNameWrongType) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -541,7 +535,7 @@ TEST(XdsBootstrapTest, CertificateProvidersUnrecognizedPluginName) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -584,7 +578,7 @@ TEST(XdsBootstrapTest, AuthorityXdsServerInvalidResourceTemplate) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -618,7 +612,7 @@ TEST(XdsBootstrapTest, AuthorityXdsServerMissingServerUri) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(
@@ -703,7 +697,7 @@ TEST(XdsBootstrapTest, CertificateProvidersFakePluginParsingError) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   EXPECT_THAT(grpc_error_std_string(error),
@@ -735,7 +729,7 @@ TEST(XdsBootstrapTest, CertificateProvidersFakePluginParsingSuccess) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   ASSERT_TRUE(GRPC_ERROR_IS_NONE(error)) << grpc_error_std_string(error);
@@ -770,7 +764,7 @@ TEST(XdsBootstrapTest, CertificateProvidersFakePluginEmptyConfig) {
   auto json = Json::Parse(json_str);
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
-  XdsBootstrap bootstrap(
+  GrpcXdsBootstrap bootstrap(
       std::move(*json),
       absl::make_unique<GrpcXdsCertificateProviderPluginMap>(), &error);
   ASSERT_TRUE(GRPC_ERROR_IS_NONE(error)) << grpc_error_std_string(error);
@@ -804,11 +798,11 @@ TEST(XdsBootstrapTest, XdsServerToJsonAndParse) {
   ASSERT_TRUE(json.ok()) << json.status();
   grpc_error_handle error = GRPC_ERROR_NONE;
   XdsBootstrap::XdsServer xds_server =
-      XdsBootstrap::XdsServer::Parse(*json, &error);
+      GrpcXdsBootstrap::XdsServerParse(*json, &error);
   ASSERT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
-  Json::Object output = xds_server.ToJson();
+  Json::Object output = GrpcXdsBootstrap::XdsServerToJson(xds_server);
   XdsBootstrap::XdsServer output_xds_server =
-      XdsBootstrap::XdsServer::Parse(output, &error);
+      GrpcXdsBootstrap::XdsServerParse(output, &error);
   ASSERT_EQ(error, GRPC_ERROR_NONE) << grpc_error_std_string(error);
   gpr_unsetenv("GRPC_EXPERIMENTAL_XDS_FEDERATION");
 }
