+++ tools/doxygen/Doxyfile.core.internal
@@ -1643,6 +1643,8 @@ src/core/ext/xds/xds_api.cc \
 src/core/ext/xds/xds_api.h \
 src/core/ext/xds/xds_bootstrap.cc \
 src/core/ext/xds/xds_bootstrap.h \
+src/core/ext/xds/xds_bootstrap_grpc.cc \
+src/core/ext/xds/xds_bootstrap_grpc.h \
 src/core/ext/xds/xds_certificate_provider.cc \
 src/core/ext/xds/xds_certificate_provider.h \
 src/core/ext/xds/xds_channel_args.h \
